+++ web/src/components/SearchField.tsx
@@ -8,9 +8,11 @@ import {
   Input,
   InputGroup,
   InputRightElement,
+  useColorModeValue,
 } from "@chakra-ui/react";
 import { useField } from "formik";
 import React, { InputHTMLAttributes, useState } from "react";
+import theme from "../theme";
 
 type SearchFieldProps = InputHTMLAttributes<HTMLInputElement> & {
   label?: string;
@@ -30,7 +32,7 @@ export const SearchField: React.FC<SearchFieldProps> = ({
     <FormControl boxSize="100%" isInvalid={!!error}>
       <InputGroup boxSize="100%">
         <Input {...field} {...props} id={field.name} boxSize="100%"></Input>
-        <InputRightElement h="100%" w="4.1rem">
+        <InputRightElement h="100%" w="4.1rem" mr={1}>
           <Button size="xs" type="submit" colorScheme="teal" isLoading={isLoading}>
             search
           </Button>